# server/openapi/openapi.yaml
openapi: "3.1.0"
info:
  title: Spraywall API
  version: "0.1.0"
  description: API for the Spraywall climbing app
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /gyms:
    get:
      operationId: listGyms
      summary: List gyms the user belongs to
      tags: [Gym]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of gyms
          content:
            application/json:
              schema:
                type: object
                properties:
                  gyms:
                    type: array
                    items:
                      $ref: "#/components/schemas/Gym"
                required: [gyms]
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createGym
      summary: Create a gym
      tags: [Gym]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGymRequest"
      responses:
        "201":
          description: Gym created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gym"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /gyms/{gymSlug}/walls:
    get:
      operationId: listWalls
      summary: List walls in a gym
      tags: [Wall]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/gymSlug"
      responses:
        "200":
          description: List of walls
          content:
            application/json:
              schema:
                type: object
                properties:
                  walls:
                    type: array
                    items:
                      $ref: "#/components/schemas/Wall"
                required: [walls]
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createWall
      summary: Create a wall
      tags: [Wall]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/gymSlug"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateWallRequest"
      responses:
        "201":
          description: Wall created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wall"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /gyms/{gymSlug}/walls/{wallId}:
    get:
      operationId: getWall
      summary: Get wall by ID
      tags: [Wall]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/gymSlug"
        - $ref: "#/components/parameters/wallId"
      responses:
        "200":
          description: Wall detail with image and detection status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WallDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    gymSlug:
      name: gymSlug
      in: path
      required: true
      schema:
        type: string
    wallId:
      name: wallId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Gym:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        created_at:
          type: string
          format: date-time
      required: [id, name, slug]
    CreateGymRequest:
      type: object
      properties:
        name:
          type: string
      required: [name]
    Wall:
      type: object
      properties:
        id:
          type: string
          format: uuid
        gym_id:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
      required: [id, gym_id, name]
    WallDetail:
      type: object
      properties:
        wall:
          $ref: "#/components/schemas/Wall"
        image:
          oneOf:
            - $ref: "#/components/schemas/WallImage"
            - type: "null"
        detection_status:
          type: ["string", "null"]
      required: [wall]
    WallImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        wall_id:
          type: string
          format: uuid
        path:
          type: string
        created_at:
          type: string
          format: date-time
      required: [id, wall_id, path]
    CreateWallRequest:
      type: object
      properties:
        name:
          type: string
        image:
          type: string
          format: binary
      required: [name, image]
...